Rough in ping support for rawsocket
Unfortunately, it's all commented out since crossbar does not currently support rawsocket pings (despite it being in the spec) in the twisted version.
https://github.com/crossbario/crossbar/issues/1206
Because of this, if a ping is sent, the system will interpret the high byte as a part of the packet length and implode. Due to this, it causes an exception and wild loops in swampyer. We simply comment this out for now until it can be reliably used. Websocket pings work fine, however.