Tweaks to make custom transport easier
This is so we can have SSH based or any other OTP based logins where we use an alternate method authenticate.
The new transport_connect function in WAMPClient makes it much easier to modify how the ticket handling or other auth handling can be handled every time a connection is renewed.