Fix Serializers Autodiscovery
Had created some code that tries to autodiscover installed serializers. Unfortunately the obj.available() call was calling obj_name.available() causing the discovery to fail entirely with no serializers found.

Turns out that by setting up the tests to be specific, it skipped the discovery code so now we're also testing that. A bug in the test code!